Home Automation Using Android Mobile
• Here we access different appliances with our mobile phone through
internet using Wi-fi.
• The micro controller used was Raspberry Pi, with an OS Raspberry in
it.
• After installing the OS, the Micro controller acts as a CPU. So, now
Raspberry Pi is the server (“installed with python socket programming”)
and the client here is a mobile phone.
• An Android application is developed with JAVA socket programming which
works with voice based commands, which is incorporated using the Google
voice recognition (which converts voice to text) when connected to
internet and acts as a client. This app sends the commands to the
Raspberry Pi (server) and the device is thus controlled.
• The Android app made is installed on the mobile phone and we keep
authenticating for a legitimate user. Once logged into the system, you
can access the appliances.